requirements.txt
attr==0.3.2
brotli==1.0.9
ConfigParser==6.0.0
cryptography==41.0.3
detectron2 @ git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0
dl==0.1.0
dotmap==1.3.30
filelock==3.12.2
fvcore==0.1.5.post20221221
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO@beeb4c29cbfa915b7aafce9bf0a23f2208498a43
hanging_threads==2.0.5
HTMLParser==0.0.2
importlib_metadata==6.7.0
ipython==8.14.0
ipywidgets==8.0.7
Jinja2==3.1.2
keyring==24.2.0
lockfile==0.12.2
lxml==4.9.3
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
omegaconf==2.3.0
onnxruntime==1.15.1
open_clip_torch==2.20.0
opencv_python==4.7.0.72
ordereddict==1.1
pandas==2.0.3
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
Pillow==9.5.0
pipeline==0.1
protobuf==3.20.3
pyarrow==12.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pynvml==11.5.0
pyOpenSSL==23.2.0
PyYAML==6.0.1
railroad==0.5.0
scipy==1.11.1
scikit-image
timm==0.9.2
toml==0.10.2
torchmetrics==1.0.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
webdataset==0.2.48
zipp==3.15.0
jupyter
ipykernel
s3cmd